Switch from float32 to float64 for fair comparison with TileSpGEMM
Change TVAL from float to double in tile_format.h and propagate to
all kernel signatures, accumulators, and host-side arrays.

+++ src/tile_format.h
@@ -10,7 +10,7 @@
 
 #define TILE_SIZE 16
 
-using TVAL = float;
+using TVAL = double;
 
 struct TileMatrix {
     int m, n, nnz;
